Add new fft driver fft_coarse_rho
This is used to transform density on the coarse grid. The problem is
in the different number of z-columns for density and wave-function
transformation. In cuFFT the overhead of creating a batched plan
is huge and should not be done often. Other solution would be
to have two internal cufft_plan_z_ in the FFT3D class: one for
density, one for wave-functions.