Preparing to keep G-vectors fixed
The idea is to add update() methods to the relevant classes such that their
internal data-structures are updated after the change of lattice vectors or
atomic coordinates.
  * add update() to Unit_cell class
  * add update() to Simulation_context class

Next classes to revisit are DFT_ground_state, Density, Potential, K_point_set

Besides, the Cartesian coordinats of G-vectors are cached and the
methods gvec_cart() and gkvec_cart() of Gvec class are now templated over
the type of index (local or global). In case of local index the precomputed
values are ruturned.