[pybind] expose density (rho, mx, my, mz) and density matrix
Add `component_checked` to `Field4d`, which raises an exception in case of invalid
access. This is in favor for a better experience in python (e.g. not
encountering segfaults).

access by descriptors:
rho, mx, my, mz and dm

for example for reading the density matrix (returns a view)
dm = density.dm
and to store into underlying C++ object:
density.dm = v