Add a stable version of Anderson mixing
and a mini app for testing purposes. Uses repeated Gram-Schmidt to
append columns and Given's rotations to drop columns.

This is a stable version of what is called broyden1 in the code,
but that's a bit of a misnomer, since our broyden1 is Anderson aka
a member of the broyden2 family.