Prototype: compose full input dictionary
The full dictionary is composed from the actual input and the schema. The JSON schema is
traversed down and for each parameter the value is taken from the actual input dictionary, or
from the schema's default value. See compose_json() function. The idea is to have one place where the
default parameters are defined and also to store the full JSON dictionary with the complete description
of the simulation in the output.