Fix symmetry check.
This commit addresses the issue with small distortions of the lattice vectors.
If the distortion is too small, spglib fails to recognise it and finds more
lattice symmetries than actually are present (changing spglib tolerance doesn't help at
the time of writing this comment). The fix is to average the rotation matrix and its
inverse tranpose when searching for the three Euler angles that generate the rotation.