Fix inconsistency when calculating the projections between block states and hubbard orbitals.
There was an inconsistency between the declaration of the matrices
containing the projections <phi_nk|phi_i> and the way they were used
and calculated.

Others modifications include explicit summation using the m(i,j)
format instead of pointers, explicit zeroing of the occupation
matrices for each atomic level even when the atomic levels do
not contribute to the hamiltonian (outside the orthogonalization
procedure).

Update of the reference outputs again. Now the amplitude of the
forces are very similar to QE. Still the ground states between
QE and QE-sirius are different so comparing hubbard forces
contributions is still difficult.