[bug] Low symmetry lattices (#812)
This PR addresses the problem of finding symmetries of a lattice with very small distrotion of a high-symmetry phase. More error checks are intoroduces. In case of failure, the control:spglib_tolerance must be decreased. The following has been done:
* pass unit cell lattice tolerance to G-vectors to be used in find_lat_sym()
* introduce a check in find_lat_sym() to test if a set of lattice operations form a group
* find G-shells based on symmetry consideration; intoroduce more checks there
* find G-vector shell length using average value of G-vector lengths
* remove lurked obsolete code

### What has not been addressed yet
Ths most of the symmetry issues come from the ill defined lattices. Typical use case is a small distortion of the high-symmetry lattice to break the symmetry and perform lattice relaxation. In that case it can make sense to not use symmetry at all. When the distorion is small, spglib can overshoot the symmetry detection and find a higher-symmetry solution (even if spglib_tolerance is relatively small). That made us introduce the addition check in find_lat_sym(). But if you just remove the symmetries that transfrom metric tensor with some threshold error, you may end up with a set of symmetry operations that do not form a group. The search for a better solution to this problem continues.