[enhancement] introduce Spheric_function_set class (#827)
Introduce a set of spherical functions and symmetrization of the set. The first use case is the PAW Vxc and Exc functions. Now the potential is computed first, then symmetrized and only then the contribution to total energy is computed. As by-product, the PAW code itself is cleaned up and simplified a little bit.